graphql-playground-html: make all settings optional
In the context of rendering the HTML page from TypeScript, you should be able to
specify a subset of settings without needing to `RecursivePartial` or anything
like that.  The React app fills in defaults for missing elements just fine. This
makes ISettings consistent with the rest of MiddlewareOptions (other than
codeTheme and schema) where fields are optional.