Don't warn if no endpoint is specified
This is a perfectly reasonable thing to do if you are serving Playground from
the same URL that also serves GraphQL, which is what Apollo Server does. Leaving
out endpoint (current plan for Apollo Server 3) will lead to PlaygroundWrapper
in the React app choosing location.href as the endpoint... which is fine, and
arguably better than the Apollo Server 2 behavior of you needing to tell it what
URL to talk to.